+++ crates/amaru-kernel/src/cardano.rs
@@ -115,6 +115,7 @@ pub mod transaction_body;
 pub mod transaction_id;
 pub mod transaction_input;
 pub mod transaction_pointer;
+pub mod transaction_ref;
 pub mod treasury_delta;
 pub mod tx_info;
 pub mod utxos;