+++ crates/amaru-kernel/src/cardano/block.rs
@@ -15,8 +15,9 @@
 use std::collections::{BTreeMap, BTreeSet};
 
 use crate::{
-    AuxiliaryData, Hash, Hasher, Header, HeaderHash, Point, Tip, Transaction, TransactionBody, WitnessSet, cbor,
-    cbor::WithSize,
+    AuxiliaryData, Hash, Hasher, Header, HeaderHash, Point, Tip, Transaction, TransactionBody, WitnessSet,
+    cardano::transaction_ref::TransactionRef,
+    cbor::{self, WithSize},
     size::{BLOCK_BODY, HEADER},
 };
 
@@ -126,6 +127,39 @@ impl IntoIterator for Block {
             })
             .collect::<Vec<_>>()
             .into_iter()
+    }
+}
+
+impl<'a> IntoIterator for &'a Block {
+    type Item = (u32, TransactionRef<'a>, u64);
+    type IntoIter = Box<dyn Iterator<Item = Self::Item> + 'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        Box::new((0u32..).zip(self.transaction_bodies.iter()).zip(&self.transaction_witnesses).map(
+            |((i, body), witnesses)| {
+                let is_expected_valid =
+                    !self.invalid_transactions.as_ref().map(|set| set.contains(&i)).unwrap_or(false);
+
+                let (auxiliary_data_len, auxiliary_data) = match self.auxiliary_data.get(&i) {
+                    Some(auxiliary_data) => (auxiliary_data.len(), Some(auxiliary_data)),
+                    None => (1, None),
+                };
+
+                // NOTE: Transaction size calculation
+                //
+                // Due to how the transactions are serialised in blocks (with seggregated witnesses
+                // and auxiliary data), we have to calculate the size from multiple pieces and add
+                // an extra 'cbor framing byte' which corresponds to the declaration of the
+                // top-level array of size 3 (`0x83`). Importantly, the validity of the transaction
+                // is not taken into account for the size calculation (rationale being that this
+                // the logic is then preserved between pre-alonzo and post-alonzo eras).
+                //
+                // See also: <https://github.com/IntersectMBO/cardano-ledger/blob/0cfbf861cfb456660a7b73281c6fb714a53d40f9/eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Tx.hs#L351-L362>
+                let size = 1 + body.len() + witnesses.len() as u64 + auxiliary_data_len;
+
+                (i, TransactionRef { body, witnesses: witnesses.as_ref(), auxiliary_data, is_expected_valid }, size)
+            },
+        ))
     }
 }